requirements.txt
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==3.2.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.3
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.30
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.8.1
